Background image added to index.html in _site
I figured the issue Pranesh was facing. I too faced the same, so I inspected the link tag that links our style.css.
Apparently, the rel attribute was missing which was causing the issue. I also changed the min-height in CSS to 80vh because it was creating additional scrolling area which is empty.